Finished the functions for accessing database
I have made some edits to the "emulator" of the vending machine to be able to access the data from the database. I will continue tomorrow to clean the code and write the comments for the code. **If you would like to try it out already download the mysql jar file and import it to the project. Try the VendingMachineEmulator class**